+++ include/exiv2/iptc.hpp
@@ -136,9 +136,9 @@ namespace Exiv2 {
         uint16_t tag() const override;
         TypeId typeId() const override;
         const char* typeName() const override;
-        long typeSize() const override;
-        long count() const override;
-        long size() const override;
+        size_t typeSize() const override;
+        size_t count() const override;
+        size_t size() const override;
         std::string toString() const override;
         std::string toString(long n) const override;
         long toLong(long n =0) const override;
@@ -293,11 +293,7 @@ namespace Exiv2 {
           @return 0 if successful;<BR>
                   5 if the binary IPTC data is invalid or corrupt
          */
-        static int decode(
-                  IptcData& iptcData,
-            const byte*     pData,
-                  uint32_t  size
-        );
+        static int decode(IptcData& iptcData, const byte* pData, size_t size);
         /*!
           @brief Encode the IPTC datasets from \em iptcData to a binary
                  representation in IPTC IIM4 format.
@@ -307,9 +303,7 @@ namespace Exiv2 {
 
           @return Data buffer containing the binary IPTC data in IPTC IIM4 format.
          */
-        static DataBuf encode(
-            const IptcData& iptcData
-        );
+        static DataBuf encode(const IptcData& iptcData);
 
     private:
         // Constant data
