+++ samples/tiff-test.cpp
@@ -48,7 +48,7 @@ void mini1(const char* path)
 
     // Write nothing, this time with a previous binary image
     DataBuf buf = readFile(path);
-    wm = ExifParser::encode(blob, buf.pData_, buf.size_, bigEndian, exifData);
+    wm = ExifParser::encode(blob, buf.pData_, (uint32_t)buf.size_, bigEndian, exifData);
     enforce(wm == wmIntrusive, Exiv2::kerErrorMessage, "encode returned an unexpected value");
     assert(blob.size() == 0);
     std::cout << "Test 2: Writing empty Exif data with original binary data: ok.\n";
